Refactoring: new file for RiskLevel changed tests
